feat: add SIMD DCT/IDCT using wide crate
Add SIMD-optimized DCT and IDCT transforms using the `wide` crate for
portable cross-platform SIMD (SSE2/NEON/WASM). The SIMD implementations
provide ~6% speedup on the full encode benchmark with native CPU:

- Scalar: 106.70ms per iteration
- SIMD: 100.26ms per iteration (6% faster)

Micro-benchmarks with -Ctarget-cpu=native show:
- DCT: 37.81ns scalar vs 34.73ns SIMD (8% faster)
- IDCT: 78.67ns scalar vs 66.67ns SIMD (15% faster)

Changes:
- Add `wide` 1.1.1 as optional dependency with new `simd` feature
- Create transform_simd.rs with dct4x4_simd, dct4x4_two, idct4x4_simd
- Delegate dct4x4() and idct4x4() to SIMD when feature enabled
- Add SIMD to default features
- Add allow(dead_code) for various pre-existing unused functions
- Fix unnecessary cast in vp8_common.rs

+++ src/vp8_common.rs
@@ -734,6 +734,7 @@ pub(crate) struct Segment {
     pub(crate) loopfilter_level: i8,
 
     /// The quantization index (0-127) for this segment
+    #[allow(dead_code)] // Stored for debugging/info purposes
     pub(crate) quant_index: u8,
 
     // Quantization matrices for trellis optimization
@@ -786,9 +787,9 @@ impl Segment {
         //
         // q values come from ExpandMatrix which returns the average of the
         // 16-element quantization matrix: (q_dc + 15*q_ac + 8) >> 4
-        let q_i4 = ((self.ydc as u32 + 15 * self.yac as u32 + 8) >> 4) as u32;
-        let q_i16 = ((self.y2dc as u32 + 15 * self.y2ac as u32 + 8) >> 4) as u32;
-        let q_uv = ((self.uvdc as u32 + 15 * self.uvac as u32 + 8) >> 4) as u32;
+        let q_i4 = (self.ydc as u32 + 15 * self.yac as u32 + 8) >> 4;
+        let q_i16 = (self.y2dc as u32 + 15 * self.y2ac as u32 + 8) >> 4;
+        let q_uv = (self.uvdc as u32 + 15 * self.uvac as u32 + 8) >> 4;
 
         self.lambda_trellis_i4 = ((7 * q_i4 * q_i4) >> 3).max(1);
         self.lambda_trellis_i16 = ((q_i16 * q_i16) >> 2).max(1);
